requirements.txt
alt-profanity-check
openai
guardrails-ai
python-dotenv 